Minor code re-org
